a bit more cleanup
